// src/types/crm.ts
export type RoleType =
  | "owner"
  | "tenant"
  | "subtenant"
  | "beirat"
  | "proxy"
  | "service_provider"
  | "caretaker"
  | "other";